+++ CHANGELOG.md
@@ -7,7 +7,8 @@ Release notes for the Router Craft CMS plugin.
 ## Unreleased
 
 ### Added
-- Added new Types filter for filtering by multiple [entry types](https://craftcms.com/docs/5.x/reference/element-types/entries.html#entry-types).
+- Added new `types` filter for filtering by multiple [entry types](https://craftcms.com/docs/5.x/reference/element-types/entries.html#entry-types).
+- Added `assumeUniqueHandles` optional parameter for the `type` and `types` filters.